feat(Table): add noHeader prop
and refactor stories to use object rather than functions
